remove: TradeFindDao
TradeService에 기능 옮김
